test: skip ':restart' spec for windows

+++ test/functional/terminal/tui_spec.lua
@@ -160,6 +160,10 @@ describe('TUI :detach', function()
   end)
 end)
 
+if t.skip(is_os('win')) then
+  return
+end
+
 describe('TUI :restart', function()
   before_each(function()
     os.remove(testlog)
@@ -214,10 +218,6 @@ describe('TUI :restart', function()
     }
   end)
 end)
-
-if t.skip(is_os('win')) then
-  return
-end
 
 describe('TUI', function()
   local screen --[[@type test.functional.ui.screen]]
